added req and res editor
