+++ src/components/code-editor/readme.md
@@ -1,4 +1,4 @@
-# code-editor
+# editor-container
 
 
 
@@ -7,31 +7,25 @@
 
 ## Properties
 
-| Property              | Attribute     | Description | Type                | Default     |
-| --------------------- | ------------- | ----------- | ------------------- | ----------- |
-| `doc`                 | `doc`         |             | `any`               | `undefined` |
-| `header`              | `header`      |             | `string`            | `undefined` |
-| `isEditable`          | `is-editable` |             | `"false" \| "true"` | `'true'`    |
-| `language`            | `language`    |             | `"java" \| "json"`  | `'java'`    |
-| `runBtn` _(required)_ | `run-btn`     |             | `"off" \| "on"`     | `undefined` |
-| `url`                 | `url`         |             | `string`            | `undefined` |
+| Property     | Attribute     | Description | Type     | Default     |
+| ------------ | ------------- | ----------- | -------- | ----------- |
+| `headerText` | `header-text` |             | `string` | `undefined` |
+| `url`        | `url`         |             | `string` | `undefined` |
 
 
 ## Dependencies
 
-### Used by
-
- - [editor-container](../editor-container)
-
 ### Depends on
 
-- [res-editor](../res-editor)
+- [fluid-container](../fluid-container)
+- [req-editor](../req-editor)
 
 ### Graph
 ```mermaid
 graph TD;
-  code-editor --> res-editor
-  editor-container --> code-editor
+  code-editor --> fluid-container
+  code-editor --> req-editor
+  req-editor --> res-editor
   style code-editor fill:#f9f,stroke:#333,stroke-width:4px
 ```
