+++ src/components/fluid-container/readme.md
@@ -16,13 +16,13 @@
 
 ### Used by
 
- - [editor-container](../editor-container)
+ - [code-editor](../code-editor)
  - [nav-bar](../nav-bar)
 
 ### Graph
 ```mermaid
 graph TD;
-  editor-container --> fluid-container
+  code-editor --> fluid-container
   nav-bar --> fluid-container
   style fluid-container fill:#f9f,stroke:#333,stroke-width:4px
 ```
