+++ src/components/req-editor/readme.md
@@ -1,4 +1,4 @@
-# editor-container
+# code-editor
 
 
 
@@ -9,24 +9,27 @@
 
 | Property | Attribute | Description | Type     | Default     |
 | -------- | --------- | ----------- | -------- | ----------- |
+| `doc`    | `doc`     |             | `any`    | `'\n\n'`    |
 | `header` | `header`  |             | `string` | `undefined` |
 | `url`    | `url`     |             | `string` | `undefined` |
 
 
 ## Dependencies
 
+### Used by
+
+ - [code-editor](../code-editor)
+
 ### Depends on
 
-- [fluid-container](../fluid-container)
-- [req-editor](../req-editor)
+- [res-editor](../res-editor)
 
 ### Graph
 ```mermaid
 graph TD;
-  editor-container --> fluid-container
-  editor-container --> req-editor
   req-editor --> res-editor
-  style editor-container fill:#f9f,stroke:#333,stroke-width:4px
+  code-editor --> req-editor
+  style req-editor fill:#f9f,stroke:#333,stroke-width:4px
 ```
 
 ----------------------------------------------
