+++ src/index.html
@@ -10,9 +10,6 @@
     <script nomodule src="build/neptune.js"></script>
   </head>
   <body class="selection:bg-indigo-200 selection:text-violet-800">
-    <!-- <main-component></main-component> -->
-    <!-- <editor-container></editor-container> -->
-    <!-- <test-com>hi</test-com> -->
-    <sign-in></sign-in>
+    <code-editor header-text="Write your Gremlin Query Here" url="http://localhost:5000/query"></code-editor>
   </body>
 </html>
